Consolidate swift build scripts into a single, better documented package
refs: none
affects: none
release note: none

Update headers and header-updating script
