+++ Canvas.xcworkspace/contents.xcworkspacedata
@@ -57,9 +57,6 @@
       </FileRef>
    </Group>
    <FileRef
-      location = "group:scripts">
-   </FileRef>
-   <FileRef
       location = "group:Student/Student.xcodeproj">
    </FileRef>
    <FileRef
