+++ scripts/build_automation/bitrise_workflows/tests.yml
@@ -66,7 +66,8 @@ workflows:
             brew tap mxcl/made
             brew tap thii/xcbeautify https://github.com/thii/xcbeautify.git
 
-            brew reinstall swiftlint jq xcbeautify mxcl/made/swift-sh
+            brew reinstall swiftlint jq xcbeautify
+
             pip3 install awscli
             rm -rf "$BITRISE_SOURCE_DIR/Pods/Target Support Files/yoga"
     - script:
@@ -76,7 +77,7 @@ workflows:
         - runner_bin: "/bin/zsh"
         - content: |-
             set -euo pipefail
-            ./scripts/build_automation/build-link.swift generate-temp-links $BITRISE_PULL_REQUEST
+            ./scripts/run-swift build-link generate-temp-links $BITRISE_PULL_REQUEST
     - &yarn-install-root
       yarn:
         title: yarn install root dependencies
@@ -177,7 +178,7 @@ workflows:
             #!/bin/zsh
             set -euo pipefail
 
-            ./scripts/build_automation/summarize-test-results.swift scripts/coverage/citests.xcresult 2>/dev/null |
+            ./scripts/run-swift summarize-test-results scripts/coverage/citests.xcresult 2>/dev/null |
                 jq -ac '{"sourcetype": "mobile-ios-testresult", "event": .}' | gzip |
                 curl -q -X POST \
                     -H "Authorization: $SPLUNK_SECRET" \
@@ -198,7 +199,7 @@ workflows:
         inputs:
           - content: |-
               set -euo pipefail
-              ./scripts/delete-extra-checkmarx-comments.swift $BITRISE_PULL_REQUEST
+              ./scripts/run-swift delete-cx $BITRISE_PULL_REQUEST
     - deploy-to-bitrise-io: {}
 
   danger-yarn:
@@ -217,7 +218,7 @@ workflows:
             mkdir -p tmp
             if ! ./scripts/runSwiftLint.sh > tmp/swift-lint-results-raw.txt 2>&1; then
                 xcbeautify < tmp/swift-lint-results-raw.txt > tmp/swift-lint-results.txt
-                ./scripts/build_automation/suggest-lint-fix.swift
+                ./scripts/run-swift suggest-lint-fix $BITRISE_PULL_REQUEST
                 exit 1
             fi
     - *yarn-install-root
@@ -525,7 +526,7 @@ workflows:
             rm -rf scripts/coverage/citests.xcresult
             mv ui-test-results/merged.xcresult scripts/coverage/citests.xcresult
             yarn coverage
-            ./scripts/coverage/export-coverage.swift
+            ./scripts/run-swift export-coverage
     - script:
         title: Report tests
         run_if: and .IsCI (enveq "BITRISE_GIT_BRANCH" "master")
@@ -535,7 +536,7 @@ workflows:
             #!/bin/zsh
             set -euo pipefail
 
-            ./scripts/build_automation/summarize-test-results.swift scripts/coverage/citests.xcresult 2>/dev/null |
+            ./scripts/run-swift summarize-test-results scripts/coverage/citests.xcresult 2>/dev/null |
                 jq -ac '{"sourcetype": "mobile-ios-testresult", "event": .}' | gzip |
                 curl -q -X POST \
                     -H "Authorization: $SPLUNK_SECRET" \
